+++ src/mpu6050sensor.cpp
@@ -173,13 +173,13 @@ double MPU6050Sensor::getAngularVelocityZ() const
   return gyro_z_converted;
 }
 
-double MPU6050Sensor::convertRawGyroscopeData(int gyro_raw) const
+double MPU6050Sensor::convertRawGyroscopeData(int16_t gyro_raw) const
 {
   const double ang_vel_in_deg_per_s = static_cast<double>(gyro_raw) / GYRO_SENS_MAP.at(gyro_range_);
   return ang_vel_in_deg_per_s;
 }
 
-double MPU6050Sensor::convertRawAccelerometerData(int accel_raw) const
+double MPU6050Sensor::convertRawAccelerometerData(int16_t accel_raw) const
 {
   const double accel_in_m_per_s =
       static_cast<double>(accel_raw) / ACCEL_SENS_MAP.at(accel_range_) * GRAVITY;
